--- web/app/components/templates/MainLayout.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { Inter } from 'next/font/google';
-import Header from '../molecules/Header';
-import '../../styles/globals.css';
-
-const inter = Inter({ subsets: ['latin'] });
-
-interface MainLayoutProps {
-  children: React.ReactNode;
-}
-
-const MainLayout: React.FC<MainLayoutProps> = ({ children }) => {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <Header />
-        {children}
-      </body>
-    </html>
-  );
-};
-
-export default MainLayout;
