+++ web/app/layout.tsx
@@ -1,16 +1,28 @@
+import React from 'react';
+import { Inter } from 'next/font/google';
+import Header from './components/molecules/Header';
+import './styles/globals.css';
+
 export const metadata = {
   title: 'Profile',
   description: 'Generated by Next.js',
 }
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+const inter = Inter({ subsets: ['latin'] });
+
+interface MainLayoutProps {
+  children: React.ReactNode;
+}
+
+const MainLayout: React.FC<MainLayoutProps> = ({ children }) => {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body className={inter.className}>
+        <Header />
+        {children}
+      </body>
     </html>
-  )
-}
+  );
+};
+
+export default MainLayout;
